Moved ArmorType enum to BaseGameObject to prepare for  other types of objects. Added ISelectable to easely tag gameobjects that can be selected.

+++ AoE/GameObjects/Units/BaseRangedUnit.cs
@@ -15,7 +15,7 @@ namespace AoE.GameObjects.Units
         
         // TODO list of projectiles
 
-        public BaseRangedUnit(Vector position, float width, float height, string name, int hitPoints, int meleeAttack, int pierceAttack, float blastRadius, float rateOfFire, int minRange, int maxRange, float accuracy, float projectileSpeed, int meleeArmor, int pierceArmor, float speed, int lineOfSight, string imageId, Team team) :
+        public BaseRangedUnit(Vector position, double width, double height, string name, int hitPoints, int meleeAttack, int pierceAttack, float blastRadius, float rateOfFire, int minRange, int maxRange, float accuracy, float projectileSpeed, int meleeArmor, int pierceArmor, float speed, int lineOfSight, string imageId, Team team) :
             base(position, width, height, name, hitPoints, meleeAttack, pierceAttack, blastRadius, rateOfFire, meleeArmor, pierceArmor, speed, lineOfSight, imageId, team)
         {
             MinRange = minRange >= 0 ? minRange : 0;
